% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{embed_contents}
\alias{embed_contents}
\title{Embed web contents from services}
\usage{
embed_contents(x)
}
\arguments{
\item{x}{request contents}
}
\value{
html character.
}
\description{
Embed web contents from services
}
\examples{
\dontrun{
embed_contents(oembed("speakerdeck",
                      url = "https://speakerdeck.com/s_uryu/nekosky"))
}
}
