% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oembed.R
\name{oembed}
\alias{oembed}
\title{Request web contents via oEmbed format}
\usage{
oembed(service, url, ...)
}
\arguments{
\item{service}{provider name}

\item{url}{a url}

\item{...}{path to httr::GET query arguments.}
}
\description{
Request web contents via oEmbed format
}
\examples{
\dontrun{
oembed("speakerdeck", url = "https://speakerdeck.com/s_uryu/nekosky")
}
}
